(#4) Reverted 'production' variable type from long to int
This is because DSP returns an int from GetItemProduction().

+++ ReusableMetadataPlugin.cs
@@ -121,7 +121,7 @@ namespace ReusableMetadata
             {
                 ClusterPropertyData clusterPropertyData = __instance.propertyDatas[i];
 
-                long production = clusterPropertyData.GetItemProduction(itemId);
+                int production = clusterPropertyData.GetItemProduction(itemId);
 
                 if (ReusableMetadataPlugin.useVerboseLogging.Value)
                 {
